Add files via upload
Model Training for Truck detection. Data set from Kaggle and additional images downloaded from internet. Dataset includes a set of images of different vehicles which are classified under the labels Truck, Car, Motorbike, Bicycle, Bus, Tractor, No Vehicle.

ImageDataGenerator: Used for rescaling images and creating training/validation splits.
Model Architecture: A simple CNN with three convolutional layers and a dense output layer corresponding to the number of classes.
Model Compilation: Uses Adam optimizer and categorical crossentropy since it's a multi-class classification problem.
Training and Validation: The model is trained with 80% of the data and validated with the remaining 20%.
Saving the Model: The trained model is saved to the specified path.